refactor(services): change name of module
